+++ leptonai/api/v1/types/deployment.py
@@ -7,7 +7,7 @@ from leptonai.config import compatible_field_validator, v2only_field_validator
 
 from .affinity import LeptonResourceAffinity
 from .common import Metadata
-from ..ingress import AuthConfig
+from .auth import AuthConfig
 
 DEFAULT_STORAGE_VOLUME_NAME = "default"
